server/requirements.txt
Flask==2.3.3
requests==2.31.0
tldextract==3.4.0
python-Levenshtein==0.23.0
python-whois==0.7.3
beautifulsoup4==4.12.2
